+++ or_globalvars.py
@@ -1,4 +1,10 @@
+import random
+import time
+
 from rich import print
+
+import or_player
+
 
 class GameGlobals:
     def __init__(self):
@@ -25,6 +31,9 @@ class GameGlobals:
         self.hostility_of_riders = False
         self.cash_total = -1
         self.has_fort = False
+        self.got_shot = False
+        self.cold_weather = False
+        self.raining = False
 
     def print_inventory(self):
         self.amount_spent_on_food = max(int(self.amount_spent_on_food), 0)
@@ -34,13 +43,13 @@ class GameGlobals:
             int(self.amount_spent_on_clothing), 0)
         self.amount_spent_on_miscellaneous = max(
             int(self.amount_spent_on_miscellaneous), 0)
-        print("=================================================")
+        print("="*53)
         print("[cyan]Food: [/cyan]", self.amount_spent_on_food)
         print("[cyan]Bullets: [/cyan]", self.amount_spent_on_bullets)
         print("[cyan]Clothing: [/cyan]", self.amount_spent_on_clothing)
         print("[cyan]Supplies: [/cyan]", self.amount_spent_on_miscellaneous)
         print("[green]Money : $[/green]", self.cash_total)
-        print("=================================================")
+        print("="*53)
 
     def increment_turn(self):
         self.current_date += 1
@@ -48,8 +57,58 @@ class GameGlobals:
     def print_too_long(self):
         print("[red]You have been on the trail for too long...[/red]")
         print("[red]Your family dies in the \
-first blizzard of the winter[/red]")
+    first blizzard of the winter[/red]")
         self.dead = True
 
     def no_turns_left(self, arr):
         return self.current_date >= len(arr)
+
+    def input_yes_no(message):
+        reply = input(message)
+        return True if 'y' in reply else False
+
+    def input_int(message):
+        text_2_int = None
+        while text_2_int == None:
+            try:
+                text_2_int = int(input(message))
+            except:
+                text_2_int = None
+        return text_2_int
+
+    def shooting():
+        words = ["BANG", "BLAM", "POW", "WHAM"]
+        word = random.choice(words)
+        t0 = time.time()
+        typed_word = input("{}".format(word))
+        t1 = time.time()
+        B1 = (t1-t0)-(vars.shooting_level)
+        if typed_word != word:
+            return 9
+        return max(B1, 0)
+
+    def illness(this_vars):
+        RND = random.random()
+        if 100*RND < 10+35*(this_vars.choice_of_eating-1):
+            print("MILD ILLNESS---MEDICINE USED")
+            this_vars.total_mileage -= 5
+            this_vars.amount_spent_on_miscellaneous -= 2
+        elif 100*RND < 100-(40/4**(this_vars.choice_of_eating-1)):
+            print("BAD ILLNESS---MEDICINE USED")
+            this_vars.total_mileage -= 5
+            this_vars.amount_spent_on_miscellaneous -= 5
+        else:
+            print("SERIOUS ILLNESS")
+            print("YOU MUST STOP FOR MEDICAL ATTENTION")
+            this_vars.amount_spent_on_miscellaneous -= 10
+            this_vars.has_illness = True
+
+
+def health():
+    if vars.is_injured == True:
+        health = "Injured"
+    elif vars.is_injured == False:
+        health = "Good"
+
+
+vars = GameGlobals()